feat: enshrine temporal as the only workflow engine (launch path + workers)
- launchPublishedWorkflowRun always launches on Temporal; engine flag and
  dead execute param removed
- startRun writes engine='temporal' unconditionally; StartRunParams loses
  its engine field
- producer and worker both use the contract task-queue constant; env
  override removed
- worker service always starts the Temporal worker; both DB-polling worker
  classes deleted
- workflow_data_store expiry sweep extracted to WorkflowDataStoreSweepWorker
