+++ services/workflow-worker/src/index.ts
@@ -11,7 +11,7 @@ import dotenv from 'dotenv';
 dotenv.config();
 
 import { initializeWorkflowRuntimeV2, registerWorkflowEmailProvider } from '@alga-psa/workflows/runtime/worker';
-import { WorkflowRuntimeV2Worker } from '@alga-psa/workflows/workers';
+import { WorkflowDataStoreSweepWorker } from '@alga-psa/workflows/workers';
 import { WorkflowRuntimeV2EventStreamWorker } from './v2/WorkflowRuntimeV2EventStreamWorker.js';
 import { WorkflowRuntimeV2TemporalWorker } from './v2/WorkflowRuntimeV2TemporalWorker.js';
 import logger from '@alga-psa/core/logger';
@@ -53,43 +53,26 @@ async function startServices() {
 
     await registerEnterpriseStorageProviders();
 
-    const runtimeV2WorkerId = `runtime-v2-${Date.now()}`;
     const runtimeV2EventWorkerId = `runtime-v2-events-${Date.now()}`;
     const runtimeV2TemporalWorkerId = `runtime-v2-temporal-${Date.now()}`;
-    const runtimeV2Worker = new WorkflowRuntimeV2Worker(runtimeV2WorkerId);
+    const dataStoreSweepWorkerId = `data-store-sweep-${Date.now()}`;
     const runtimeV2EventWorker = new WorkflowRuntimeV2EventStreamWorker(runtimeV2EventWorkerId);
     const runtimeV2TemporalWorker = new WorkflowRuntimeV2TemporalWorker(runtimeV2TemporalWorkerId);
-    const enableDbPollingWorker =
-      process.env.WORKFLOW_RUNTIME_V2_ENABLE_DB_POLLING === 'true'
-      || process.env.WORKFLOW_RUNTIME_V2_ENABLE_DB_POLLING === '1';
-    const enableTemporalPolling = !['false', '0', 'no'].includes(
-      String(process.env.WORKFLOW_RUNTIME_V2_ENABLE_TEMPORAL_POLLING ?? 'true').toLowerCase(),
-    );
+    const dataStoreSweepWorker = new WorkflowDataStoreSweepWorker(dataStoreSweepWorkerId);
     logger.info('[WorkflowWorker] Starting runtime v2 workers', {
-      runtimeV2WorkerId,
       runtimeV2EventWorkerId,
       runtimeV2TemporalWorkerId,
+      dataStoreSweepWorkerId,
       consumerGroup: process.env.WORKFLOW_RUNTIME_V2_EVENT_CONSUMER_GROUP ?? 'workflow-runtime-v2',
-      temporalQueue: process.env.WORKFLOW_RUNTIME_V2_TEMPORAL_TASK_QUEUE ?? 'workflow-runtime-v2',
       temporalAddress: process.env.TEMPORAL_ADDRESS ?? 'temporal-frontend.temporal.svc.cluster.local:7233',
       temporalNamespace: process.env.TEMPORAL_NAMESPACE ?? 'default',
-      enableDbPollingWorker,
-      enableTemporalPolling,
     });
-    if (enableTemporalPolling) {
-      await runtimeV2TemporalWorker.start();
-      healthServer.setWorker('temporal', true);
-    } else {
-      logger.info('[WorkflowWorker] Authored Temporal polling disabled for this environment');
-    }
+    await runtimeV2TemporalWorker.start();
+    healthServer.setWorker('temporal', true);
     await runtimeV2EventWorker.start();
     healthServer.setWorker('eventStream', true);
-    if (enableDbPollingWorker) {
-      await runtimeV2Worker.start();
-      healthServer.setWorker('dbPolling', true);
-    } else {
-      logger.info('[WorkflowWorker] DB polling runtime worker disabled; Temporal-native runtime is authoritative for new runs');
-    }
+    await dataStoreSweepWorker.start();
+    healthServer.setWorker('dataStoreSweep', true);
 
     healthServer.markReady();
     logger.info('[WorkflowWorker] All services started successfully');
@@ -101,9 +84,9 @@ async function startServices() {
     async function shutdown() {
       logger.info('[WorkflowWorker] Shutting down services...');
       await Promise.all([
-        enableTemporalPolling ? runtimeV2TemporalWorker.stop() : Promise.resolve(),
+        runtimeV2TemporalWorker.stop(),
         runtimeV2EventWorker.stop(),
-        enableDbPollingWorker ? runtimeV2Worker.stop() : Promise.resolve(),
+        dataStoreSweepWorker.stop(),
         healthServer.stop(),
       ]);
       process.exit(0);
